+++ src/tsp/src/tsp_driver.cpp
@@ -49,7 +49,7 @@ typedef std::vector< int64_t > Ids;
 int
 do_pgr_tsp(
         Matrix_cell_t *distances,
-        int64_t total_distances,
+        size_t total_distances,
         int64_t start_vid,
         int64_t end_vid,
         General_path_element_t **path,
