+++ src/tsp/src/tsp_driver.h
@@ -38,8 +38,7 @@ extern "C" {
 
 int 
 do_pgr_tsp(
-        Matrix_cell_t *distances,
-        int64_t total_distances,
+        Matrix_cell_t *distances, size_t total_distances,
         int64_t start_vid,
         int64_t end_vid,
         General_path_element_t **results,
